Crate aws_runtime

Runtime support code for the AWS SDK. This crate isn’t intended to be used directly.

Modules

auth

Supporting code for authentication in the AWS SDK.

content_encoding

AWS-specific content-encoding tools

env_config

Supporting code for parsing AWS config values set in a user’s environment or in a shared config file.

fs_util

Filesystem utilities

invocation_id

Supporting code for invocation ID headers in the AWS SDK.

recursion_detection

Supporting code for recursion detection in the AWS SDK.

request_info

Supporting code for request metadata headers in the AWS SDK.

retries

Supporting code for retry behavior specific to the AWS SDK.

service_clock_skew

Interceptor that determines the clock skew between the client and service.

user_agent

Supporting code for user agent headers in the AWS SDK.