Module aws_runtime::retries::classifiers

Classifiers that can inspect a response and determine if it should be retried.

Structs

AwsErrorCodeClassifier

A retry classifier for determining if the response sent by an AWS service requires a retry.

AwsErrorCodeClassifierBuilder

Builder for AwsErrorCodeClassifier

Constants

THROTTLING_ERRORS

AWS error codes that represent throttling errors.

TRANSIENT_ERRORS

AWS error codes that represent transient errors.