Module aws_runtime::user_agent

Supporting code for user agent headers in the AWS SDK.

Structs

ApiMetadata

Metadata about the client that’s making the call.

AwsUserAgent

AWS User Agent

InvalidMetadataValue

Error for when an user agent metadata doesn’t meet character requirements.

UserAgentInterceptor

Generates and attaches the AWS SDK’s user agent to a HTTP request

Enums

BusinessMetric

Enumerates human readable identifiers for the features tracked by metrics