aws_runtime
Module env_config

Supporting code for parsing AWS config values set in a user’s environment or in a shared config file.

Modules

error

Errors related to AWS profile config files

file

Config structs to programmatically customize the profile files that get loaded

parse

Profile file parsing

property

Sections within an AWS config profile.

section

Sections within an AWS config profile.

source

Code for handling in-memory sources of profile data

Structs

EnvConfigError

An error occurred when resolving config from a user’s environment.

EnvConfigSource

The source that env config was derived from.

EnvConfigValue

Environment config values are config values sourced from a user’s environment variables or profile file.

Functions

get_service_env_config

Given a key, access to the environment, and a validator, return a config value if one was set.