Struct aws_runtime::recursion_detection::RecursionDetectionInterceptor

#[non_exhaustive]
pub struct RecursionDetectionInterceptor { /* private fields */ }

Recursion Detection Interceptor

This interceptor inspects the value of the AWS_LAMBDA_FUNCTION_NAME and _X_AMZN_TRACE_ID environment variables to detect if the request is being invoked in a Lambda function. If it is, the X-Amzn-Trace-Id header will be set. This enables downstream services to prevent accidentally infinitely recursive invocations spawned from Lambda.

Implementations

impl RecursionDetectionInterceptor

pub fn new() -> Self

Creates a new RecursionDetectionInterceptor